+++ test/test.js
@@ -3,5 +3,5 @@ require ('./components/extractIdTest.js');
 require ('./components/isValidIdTest.js');
 require ('./components/loadJSONTest.js');
 require ('./components/simplifyGeometryTest.js');
-require ('./components/filterOutDeprecatedRecordsTest.js');
+require ('./components/isActiveRecordTest.js');
 require ('./components/filterOutNamelessRecordsTest.js');
